optimization/new rotation
I want to save all the performace we can so that the pathfinding will
work better. I made most of the animation stuff work similarly, but
without loading a new picture every couple frames. Also, rotation has
been fixed so we can rotate around points instead of just centers which
is nice because most of out pictures are off center.